requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.5
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pygments==2.7.4
PySocks==1.7.1
pytz==2020.5
PyYAML==5.4
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
